pyproject.toml
[project]
name = "vote-talk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
]
